trading_bot/requirements.txt
anthropic>=0.30.0
python-binance>=1.0.19
apscheduler>=3.10.4
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
loguru>=0.7.2
rich>=13.7.1
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx[http2]>=0.27.0
twikit>=2.3.3
